requirements.txt
# Environment & config
python-dotenv>=0.19.0

# Google Slides integration
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0

# AI services
openai>=1.0.0
elevenlabs>=1.57.0
pipecat-ai>=0.0.58

# PDF parsing (optional)
PyPDF2>=3.0.1

# Audio capture for STT
sounddevice>=0.4.0
soundfile>=0.10.0

# Audio playback
playsound>=1.2.2
pyobjc-framework-Cocoa>=9.0

# Additional dependencies
numpy>=1.21.0
requests>=2.26.0
aiohttp>=3.8.0
python-multipart>=0.0.5